+++ backend/src/test/java/com/allog/dallog/acceptance/SubscriptionAcceptanceTest.java
@@ -16,8 +16,8 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import com.allog.dallog.common.config.TestConfig;
 import com.allog.dallog.domain.category.dto.request.CategoryCreateRequest;
 import com.allog.dallog.domain.category.dto.response.CategoryResponse;
-import com.allog.dallog.domain.subscription.dto.SubscriptionUpdateRequest;
 import com.allog.dallog.domain.subscription.dto.request.SubscriptionCreateRequest;
+import com.allog.dallog.domain.subscription.dto.request.SubscriptionUpdateRequest;
 import com.allog.dallog.domain.subscription.dto.response.SubscriptionResponse;
 import com.allog.dallog.domain.subscription.dto.response.SubscriptionsResponse;
 import io.restassured.RestAssured;
